fix: import paths after src folder rename
