+++ src/cmd/cz.go
@@ -3,8 +3,8 @@ package cmd
 import (
 	"fmt"
 
-	"github.com/Slug-Boi/cocommit/src_code/go_src/cmd/tui"
-	"github.com/Slug-Boi/cocommit/src_code/go_src/cmd/utils"
+	"github.com/Slug-Boi/cocommit/src/cmd/tui"
+	"github.com/Slug-Boi/cocommit/src/cmd/utils"
 	"github.com/inancgumus/screen"
 	"github.com/spf13/cobra"
 )
