+++ src/cmd/root.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/Slug-Boi/cocommit/src_code/go_src/cmd/tui"
-	"github.com/Slug-Boi/cocommit/src_code/go_src/cmd/utils"
+	"github.com/Slug-Boi/cocommit/src/cmd/tui"
+	"github.com/Slug-Boi/cocommit/src/cmd/utils"
 
 	"github.com/inancgumus/screen"
 	"github.com/spf13/cobra"
