+++ src/cmd/tui/tui_list.go
@@ -2,7 +2,7 @@ package tui
 
 import (
 	"fmt"
-	"github.com/Slug-Boi/cocommit/src_code/go_src/cmd/utils"
+	"github.com/Slug-Boi/cocommit/src/cmd/utils"
 	"io"
 	"os"
 	"sort"
