+++ src/main.go
@@ -4,7 +4,7 @@ Copyright © 2024 NAME HERE <EMAIL ADDRESS>
 */
 package main
 
-import "github.com/Slug-Boi/cocommit/src_code/go_src/cmd"
+import "github.com/Slug-Boi/cocommit/src/cmd"
 
 func main() {
 	cmd.Execute()
